Delete host risk card from overview page
