+++ x-pack/plugins/security_solution/public/hosts/components/kpi_hosts/index.tsx
@@ -12,9 +12,9 @@ import { HostsKpiHosts } from './hosts';
 import { HostsKpiUniqueIps } from './unique_ips';
 import type { HostsKpiProps } from './types';
 import { CallOutSwitcher } from '../../../common/components/callouts';
-import { RISKY_HOSTS_DOC_LINK } from '../../../overview/components/overview_risky_host_links/risky_hosts_disabled_module';
 import * as i18n from './translations';
 import { useHostRiskScore } from '../../../risk_score/containers';
+import { RISKY_HOSTS_DOC_LINK } from '../../../../common/constants';
 
 export const HostsKpiComponent = React.memo<HostsKpiProps>(
   ({ filterQuery, from, indexNames, to, setQuery, skip, updateDateRange }) => {
