+++ x-pack/plugins/security_solution/public/overview/components/link_panel/inner_link_panel.tsx
@@ -8,7 +8,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { EuiFlexGroup, EuiFlexItem, EuiIcon, EuiLink, EuiSplitPanel, EuiText } from '@elastic/eui';
-import { LEARN_MORE } from '../overview_risky_host_links/translations';
+import * as i18n from './translations';
 
 const ButtonContainer = styled(EuiFlexGroup)`
   padding: ${({ theme }) => theme.eui.euiSizeS};
@@ -66,7 +66,7 @@ export const InnerLinkPanel = ({
               data-test-subj={`${dataTestSubj}-learn-more`}
               external
             >
-              {LEARN_MORE}
+              {i18n.LEARN_MORE}
             </EuiLink>
           )}
         </p>
