+++ x-pack/plugins/security_solution/public/overview/components/link_panel/translations.ts
@@ -1,0 +1,15 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { i18n } from '@kbn/i18n';
+
+export const LEARN_MORE = i18n.translate(
+  'xpack.securitySolution.overview.linkPanelLearnMoreButton',
+  {
+    defaultMessage: 'Learn More',
+  }
+);
