+++ x-pack/plugins/security_solution/public/overview/pages/overview.tsx
@@ -30,9 +30,7 @@ import { useDeepEqualSelector } from '../../common/hooks/use_selector';
 import { ThreatIntelLinkPanel } from '../components/overview_cti_links';
 import { useAllTiDataSources } from '../containers/overview_cti_links/use_all_ti_data_sources';
 import { useUserPrivileges } from '../../common/components/user_privileges';
-import { RiskyHostLinks } from '../components/overview_risky_host_links';
 import { useAlertsPrivileges } from '../../detections/containers/detection_engine/alerts/use_alerts_privileges';
-import { useIsExperimentalFeatureEnabled } from '../../common/hooks/use_experimental_features';
 import { LandingPageComponent } from '../../common/components/landing_page';
 
 const OverviewComponent = () => {
@@ -68,15 +66,6 @@ const OverviewComponent = () => {
   const { hasIndexRead, hasKibanaREAD } = useAlertsPrivileges();
   const { tiDataSources: allTiDataSources, isInitiallyLoaded: isTiLoaded } = useAllTiDataSources();
 
-  const riskyHostsEnabled = useIsExperimentalFeatureEnabled('riskyHostsEnabled');
-
-  const timerange = useMemo(
-    () => ({
-      from,
-      to,
-    }),
-    [from, to]
-  );
   return (
     <>
       {indicesExist ? (
@@ -143,15 +132,6 @@ const OverviewComponent = () => {
                             from={from}
                             setQuery={setQuery}
                             to={to}
-                          />
-                        )}
-                      </EuiFlexItem>
-                      <EuiFlexItem grow={1}>
-                        {riskyHostsEnabled && (
-                          <RiskyHostLinks
-                            deleteQuery={deleteQuery}
-                            setQuery={setQuery}
-                            timerange={timerange}
                           />
                         )}
                       </EuiFlexItem>
